Add `skip_frame nokey` to timelinepreview operation
This significantly speeds up the operation. Some examples:
- 5min video h264 1080p 190MB: 5.0s -> 1.5s
- 5min video h265 2160p 540MB: 32.5s -> 3.3s
- 50min video h264 2160p 200MB: 75s -> 14s

For short videos, the timelinepreview will be less and less useful with
this: if a very short video only has 5 key frames, the generated image
will only contain 5 different frames, most repeated multiple times. For
that reason, I only enabled this for videos longer than 120s. Videos
longer than that should have enough keyframes such that only using those
shouldn't lead to a significant drop in preview quality.

However, there are fringe cases for which this change makes it worse:
videos 121s long, but with extremely long GOPs (and thus very few
keyframes). However: this won't crash for those videos, the output will
just be worse. And: I think those videos are very rare and the savings
in terms of processing time are worth it in my opinion.

Finally, ideally, in the future, this operation should use the smallest
output encoding as input (e.g. the 480p stream). The preview downscales
the frame a lot anyway. That would:
- Speed things up further, given that ffmpeg has to decode smaller
  frames.
- Make sure that we have somewhat regular keyframes to completely rule
  out fringe cases.