Fix ltitools Upload crashing if extendedMetadata exists
The ltitools Upload page can crash on loading, if the series
specified for the upload is attached to an event which
has at least one extendedMetadata dublincore catalog
attached to it (or any other dublincore catalog beside the
default one really).

This fixes that, although the fix relies on hardcoding the
default flavor of the event dublincore catalogs, which is
technically configurable. I have not found a way to easily
access whether a dublincore catalog is the common one,
and don't currently feel motivated to implement one :').