Include all metadata in Tobira harvest API for series
Before, the information returned by the harvest API was rather limited.
I singled out `created` as a separate field as it seems like it might
be more important than the others and because I want to transmit it as
timestamp (like all other dates) instead of the ISO string from the
catalog.