Reapply original patch 6196bd4444068a5a444f60cea0f667ede6345724
Whisper transcription engine bug fixing

This patch fixes a number of minor bugs and inconveniences in the
whisper transcription engine. This includes loading of default values
when the configuration has been modified which incorrectly kept their
old values in some cases. It also does proper debug logging, not
filtering out the actual output, which is what you want for debugging.

# Conflicts:
#	modules/speech-to-text-impl/src/main/java/org/opencastproject/speechtotext/impl/engine/WhisperEngine.java