Fixed created value on scheduling event series
The values of start date field and created field are synchronized. All changes of start date value will trigger an update of created value. Also on upload or schedule event. But on schedule of multible events the created value is not set correctly. It stick at first start date on all events.