Remove utility class Snapshots.java
Removes a utility class that consists of (mostly unused)
convenient shothands for use with entwine streams. Since
we are moving away from entwine streams and none of the
functions in Snapshots.java are complex enough to really
warrant an utility class, I believe nothing of values is
lost by this deletion.