Annotate classes with @JaxrsResource
This adds the osgi.jaxrs capability to the kernel bundle, which
fakes jaxrs whiteboard specification support. This allows us to
implement only the parts of the specification we use with the benefit of
up-to-date dependencies. Once a new implementation is available, we
can use it as a drop in replacement.