Move custom roles addition for episode id roles from Asset Manager to ElasticsearchIndex
The code in the Asset Manager was only adding the custom roles
required by episode id roles during index rebuilds. This is of
course not helpful when you for example have a new event that
should automatically have those custom roles.

This is why the code was moved from the Asset manager to
ElasticsearchIndex update functions. This means the code now
runs everytime an event is updated in the index. This may
not be great for performance, but is ultimately necessary?