Change Playlist Update behaviour
Changes how playlist endpoints update playlists by adding
a new method to the playlist service. Instead of removing
fields from a playlist if they were not present, this change
leaves them intact. Now, in order to remove a field from a
playlist, the field needs to be explicitely set to "null",
e.g. `"description": null`. This should make updating playlists
easier and more future proof.

Warning, this does not work for entries in arrays! If you wish
to update a single entry in an array, you still have to send the
whole array. If you only send the entry you wish to update, all
other entries will be removed.