Series xacml and extended metadata handling on ingest
This patch fixes the handling of series extended metadata and ACLs on ingest.
On ingest of a mediapackage, which includes series metadata and xacml, the existence of the series in Opencast will be checked. If the series does not exist, it will be created based on the series metadata catalog from mediapackage. If the series exists, metadata will only be updated if the ingest service configuration allows it (with default value of false). In both cases, only the series dublincore from the mediapackage is used. The series xacml and extended metadata are ignored. This patch fixes this behavior. The series xacml will be taken into account on a newly created series. Also, the extended metadata will be applied.