Fix deleting user references from the admin UI
Fixes #5753, where a `matterhoron-reference` user could not be
deleted, presumably because it wasn't found.

This was broken by #5062, and looks like a classic merge conflict
resolution mistake to me. The result is that we look up the user we
want to delete in the user and roles provider twice, and ignore the
user reference provider, which happens to be `null` anyway, since it
is never bound in the first place.