Invalidate cached user on changes
In certain circumstances, you may want to configure multiple authentication mechanisms in Opencast, but with the same user source. For example, a user should be able to log in using Shibboleth and also using JWT from an external trusted application. In this case, the user roles should be merged. This is handled correctly by Opencast. User data is then cached for performance reasons. However, on the first request using a different login mechanism, the user details should be set correctly, but the old cached user details are used instead. All subsequent requests are fine. This patch should fix that.