Improve Assetmanager multi store behavior
Since PR #4955, Assetmanager supports multiple stores. Moving a mediapackage to different store can result in an AssetManagerException like this:
`An asset with checksum … (md5) has already been archived but trying to copy or link asset StoragePath(orgId=mh_default_org, mpId=…, version=…, mpeId=…) to it failed`
This is because the path to the mediapackage is cached in the Assetmanager for 60 minutes. That's too long. Reducing the period to 1 minute is enough to get performance boost and be dynamic.

Assetmanager decides which asset store to use for each snapshot. That's not the best way. If you have a snapshot in store A and you create another snapshot, the second one may go to store B. This means that hardlinking between the two asset stores may not work and the files will be copied. This patch changes this behavior. All snapshots of a mediapackage are stored in the same asset store.