Change segment `startTime` from mpeg7 to milliseconds in Tobira API
All durations in the API are transmitted as milliseconds, so this should
be too. On the Tobira side we want to save it as milliseconds too,
since Paella wants that anyway.

This is a breaking API change and the API was already released as OC
15.5, so why no API version bump? The Tobira API is declared internal
and only to be used by Tobira. And Tobira did not use it yet! So nothing
can break.