Replace Elastic "bool" query for roles with "terms" query
The manually built "bool" query (i.e. "has this role OR has that role")
can easily bloat the query a lot. When a user has lots of roles, it is
possible to run into limits. Using the "terms" query seems more
appropriate as we can provide all user roles as array. The JSON sent to
Elastic is noticeably smaller with this and the "query limit" won't be
a problem.

The "terms" query is described as:

> Returns documents that contain one or more exact terms in a provided
> field.

For array fields, this works as expected: it is matched if there is any
overlap between the input roles and the roles in the ACL.

Before, each part of the bool query was a `matchQuery`, which can allow
for fuzzy matching. However, the ACL fields have the type `keyword` (and
not `text`), so no fuzzy analysis is done anyway. I tested this with a
few examples.