Add `SearchResult::getCreatedDate` and use it to fix Tobira harvest API
Before 16, the search service already returned a `Date` for `created`.
When we switched to ElasticSearch, this was turned into a formatted
String instead, which first went unnoticed. When noticed, it was fixed
by calling `Instant.parse` which works for some cases, but not all.

The date format returned by ES is not always the same. When adding
documents to ES, it will parse fields declared as dates and store the
date as "ms since epoch UTC" internally to be able to filter by it. But
it also remembers the original format and does not format the ms in a
well defined format. The formats that can be parsed by ES can be
configured, but by default (our case) it's
"strict_date_optional_time||epoch_millis". Documents with invalid dates
will lead to an exception, i.e. will not be added to the index. So when
querying the index, we have to expect any date format that can be parsed
by the above default format. Compare [1].

And well, that's what this PR does: use the correct method from ES to
parse it. I added the method to `SearchResult` as I figured that might
be useful in the future.

[1]: https://www.elastic.co/guide/en/elasticsearch/reference/current/date.html