Make admin-ng a binary
The old admin interface is deprecated and we don't want to maintain,
modify, fix it any longer. That means it is unnecessary to re-build the
module every time we build Opencast. Especially since building the
module takes forever.

This patch changes the mechanism to use a pre-built binary of the
current state instead. The binary is downloaded and will be included
similar to how we include new admin interface or editor releases.