Fix publication icon paths for new admin ui
The paths for the icons specified per default in
listproviders/publication.channels.properties have
changed for the new admin ui, so the icons do not
show up. This fixes that by correcting the path.

This inevitably will make the icons not show up
in the old admin ui though. I tried to mitigate that
a bit by leaving a comment about the differences.