Display state mappings in the status column again
The events table in the admin ui has a status column that informs
users about the current state of the latest workflow. A workflow
can define "state mappings" to overwrite what is usually displayed.
For example, a delete workflow may overwrite the usualy "Running"
with "Deleting". This has been broken since Opencast 12-ish, and
this commit should fix it.