No Playlists on Presentation
The playlist service requires the index service which is not present on
presentation nodes, causing the module to fail. This just removes
playlist from the node.

It also moves the Tobira module to the admin since it now requires the
new playlist module and it will break otherwise.

That's not great and ideally, the service shouldn't require the index
service but this is a quick fix to the immediate problem.