State maiing for fast workflow
The fast workflow usually processes and publishes videos. It's strange
to label the running state “Importing”. Even weirder to do that when the
workflow is failing.

Maybe this is something no one noticed since we migrated to YAML
workflows as the state mapping is broken for a while now.

In any case, a generic “Running” fits the workflow definitely better.