Fix Tobira Harvest API (includesItemsUntil & hasMore)
This was badly broken when this was changed to use the new search
service (5880526444317ea9396c97f4fa876480b83cd09d). There,
`getTotalHits()` was accidentally used, which returns the total number
of events still left. This of course basically never exactly equals
`preferredAmount + 1`. `.getHits().size()` is the correct number to use
here. However, as a precaution, I changed all `==` to `>=` just in case
something like this happens in the future again. The logic is still
sane, no need to have this exact equality.