Prevent NullPointerException in ACL parser
Parsing `{}` as ACL resulted in a NullPointerException since Opencast
just expected the object to contain a field `acl`. That means, something
like `{"acl":{}}` would be fine and result in an empty ACL. The empty
object should be handled gracefully and be treated the same. That is,
what this patch does.