Editor processing should be enabled if track selection isn't available
With track selection in the editor you want to disable processing if no tracks are selected. But if the track previews weren't generated, the user can't select any track. This will disable 'Save & Continue' button for further processing in the old editor. This behavior was added with commit 04df111 and should be soften in case of selection can't be done.

Signed-off-by: Waldemar Smirnow <wsmirnow@uni-osnabrueck.de>