Properly handle ACL publication for non-admins
If a non-admin user tries to publish something, Opencast will try to get
the ACL for the media package intended for publication. During this
process, Opencast tries to evaluate the ACL and write it to the
database. Unfortunately, this process silently fails, causing an empty
ACL to be written into the database instead.

The search service then tries to publish to the search index. This
causes several files linked in the media package to be accessed for
which the ACL in the database is being evaluated. Since that is now an
empty ACL, the access check fails, meaning that the publication fails
halfway (event is in the database, but not in the index).

This patch allows Opencast to always access the access control list
linked in the media package, so that the correct ACL ends up in the
database, meaning the correct ACL will be used for the index checks.

This fixes the overall problem and allows non-admin users to actually
publish events.

This fixes #5333
This fixes #6040