Make Tobira harvest API work with multi-tenancy
An org-filter was added to the event-query and the playlist
`getForAdministrativeRead`. The latter also relaxed its role
requirements to work with role admin roles as well. And the Tobira
endpoint itself now checks for (org-)admin role. The search query did
not need any adjustments, it already filtered by the org.