Replace AngularJS based login page
This patch replaces the AngularJS based login page. Since the code was
extremely simple, this does not need any external libraries. A few lines
of native JavaScript is sufficient to get the same functionality.

This also simplifies the CSS quite a bit. Although, you could probably
still do a better job if you redo this from ground up with modern CSS.
But this should already be much simpler.

The login page is now also somewhat responsive.

Finally, this will cause the default credentials to be pre-filled into
the login form if this is detected to be a test or development
installation. Meaning, either this run on `localhost` or on a
`opencast.org` domain.