Remove old database update test
The tests seem to be broken¹ for a while and for MySQL. Just fixing this
or removing MySQL doesn't make a lot of sense since updates work
differently by now and would need to take automatically generated
database structures into account as well. That is why this patch is
removing the old test.

¹⁾ https://github.com/opencast/opencast/actions/workflows/test-database-upgrade.yml