Add documentation on how to review pull requests
Starts with fixing #5968.

Adds documentation on what to look out for when reviewing an Opencast pull request.
Does not tell you where to click, but about all the things you could be reviewing.

Mostly a summary of @ŋregorydlogan 's talk from the Opencast Summit 2021 which is now
also linked in the documentation.