Allow Amberscript transcriptions to be attached as tracks
Configuring the `amberscript-attach-transcription` workflow operation
handler to add subtitles as tracks to the media package has no effect
and they will still be added as attachments since the SubRib converter
always returns attachments.

This patch fixes the problem by simplifying the configuration parsing,
converting the attachment to a track if necessary and also adding the
recommended generator tags to the final media package element.