Allow JWT roleMapping expressions to return list/array of strings
Previously, they could only return a single string (or null), meaning
that there was no way to have one claim result in dynamically many
roles. This patch fixes this by adding the ability to return a list or
array of strings.

Returning a single string is still allowed, so this is backwards
compatible. The new code has a bunch of casts that will throw an
exception if the type does not match, but that was already the behavior
before.