Fix integration tests, distribution upload and demo server update
We switched to Java 17 and 23 on `develop`, but the integration tests
and the distribution upload were still set to only happen for Java 11
builds.

This means that the “Test Opencast” workflow finished successfully, but
the integration tests were skipped and develop.opencast.org did not get
updated for a while now.