Add per-organization bucket name configuration
Add the ability to configure S3 buckets for each organization
(asset manager and distribution). The old configuration is still
compatible and will result in a default bucket configuration,
i.e. if no org-specific buckets are configured, the default bucket
will be used.

This likely depends on #6252 (I have not tested without this
additional patch).