Allow rolling back to a specified snapshot in asset manager
This patch improves the `asset-delete` workflow operation handlöer to
not only allow for deleting all or all except the latest snapshot of an
event, but also allows deleting all version newer than a given version.
This essentially means that you can roll back to a given event in the
asset manager.

Along the way, this also fixes the problem of this operation creating
errors since it may delete files the current workflow requires to work
on. This could previously be worked around using `ingest-download`, but
is actually quite unnecessary.