Plugins with different modules on different distributions
Our plugin mechanism already allowed for plugins to only become active
if certain Karaf features are present. This means you can have a plugin
active in the configuration of all your servers, but it only becomes
active on the admin since it only makes sense to run it there.

Unfortunately, this becomes problematic if you actually want to activate
different modules on different servers if the plugin becomes active.
This would be the usual case for things we run on the worker, where
there would still be an api, a remote and a woh module on the admin.

This patch introduces the concept of plugin groups. This means that we
can now define several plugins for different distributions like:

- opencast-plugin-xy_a
- opencast-plugin-xy_b

Which can have different activation rules, but which are all activated
if you set `opencast-plugin-xy = on` in the plugin configuration.

As an example, this patch turns the crop service into a distributed
plugin. The plugin is disabled by default since the service is not used
by default in Opencast.