Add user info to roles for admin ui
Adds user information for user roles to admin ui
GET endpoints. The frontend wants to render user
roles like ROLE_USER_ARNE with user information like
username and email, so instead of having to do perform
multiple GET requests and then mix-and-matching roles
and users in the frontend, it seemed prudent to change
the backend instead.

To that end, ACLs are not sent as is anymore, but transformed
into the format used by the frontend so that user information
can be more easily attached.