Fix multiple creation of subtitle files
Fixes the creation of multiple subtitle files, when someone clicks on "save" or "save and process" multiple times without reloading or restarting the editor.

The fix changes the initial uuid of the subtitle file. Instead of creating always a new and random one in the backend, we use the uuid provided by the frontend. So when clicking "save" multiple times, the backend code notices it's the same uuid and will not create new files every time.