Address OSGi service retrieval exception
An exception occurs during the activation of the GraphQLServlet and
SchemaServlet if their references are not yet available. This situation
arises due to the reactivation of the referenced components during
configuration updates when no modified method is specified.

To resolve this issue, an explicit modified method has been implemented
for both the ExecutionService and the SchemaService. This enhancement
allows configuration updates to be processed without the need to
reactivate the component.

Signed-off-by: Gregor Eichelberger <gregor.eichelberger@tuwien.ac.at>