Respect order of sort parameters for search queries
Fixes an issue where the sort order of the results of an elasticsearch query would be
wrong if multiple sort parameters were specified.
Since the order of the sort parameters in a query matters, our code should respect the
ordering of sort parameters. Hence I switched the unordered collection we were using
for an ordered one.

### How to test this

Can for example be tested with a query like this: admin-ng/event/events.json?limit=10&offset=0&sort=title:DESC,uid:asc