Log ID of event failing in scheduler index rebuild
We had a few problems with rebuilding the scheduler index.
Unfortunately, the logs just told us that the rebuild failed, not which
event failed. That made debugging the error really hard.

This patch adds the media package identifier to the error message, to
avoid the same issue in the future.