Log Publications Blocking Event Deletion
If you click on “Delete” in the admin interface, Opencast will try to
delete the given event, but refuse to actually delete it if not all
publications are retracted as part of the started workflow.

Unfortunately, Opencast does not tell you anything about what
publications are blocking the deletion. That's annoying if you want to
debug why an event is not being deleted.

This patch adds a list of publications blocking the deletion to the
logs.