Don't fail removal on already deleted workflow
If you try to delete an event, but one of it's workflows has no job in
the database because it has already been deleted for whatever reason,
the whole process will fail. Since we want it to be deleted, this should
just acknowledge that the event is already gone and continue.