Improve Performance of Metrics Endpoint
This patch drastically improves the performance of the metrics endpoint.
The metrics gathering was one factor drastically worsen our already
crippled system when our Opencast cluster broke down after upgrading to
the Stud.IP Opencast plugin 3.x two weeks ago.

The most important change included in this patch is moving to a native
SQL query for counting active jobs and workflows grouped by
organizations and (in case of jobs) hosts. Run on a snapshot of our
post-crash production database, this took the time to calculate the
values from about 20 seconds down to about 150 milliseconds. This should
be no more problem. Even with an escalating number of jobs.

This also uses way less memory and CPU cycles on the Opencast nodes
since much of the actual performance was lost by deserializing the
database result and turning it into Java objects.

To be fair, the old code is no problem for normal operations. The code
makes the assumption that you operate on a reasonable amount of jobs.
The problem was that this was no longer true for our already struggling
system.

A second improvement is the removal of organizational filter on a
single-tenancy system. The organization filter can be a somewhat
expensive operation as well. On my test system, this additional change
speed things up by an additional 14 seconds, dropping the overall
execution time to less than a second.

I suspect, however, that this slow query is partly due to resource
limitations on my test system. It is only a .8 second difference on our
somewhat optimized production database. Still, it's a simple change to
gain more performance.