Add isOnline to admin ui service endpoint
Makes the "GET /services.json" endpoint return additional information
on whether a service is online, offline or in maintenance. This can then
be displayed in the admin ui to make it easier to discern whether
the service is online, without having to first match it with its host
in the servers table.