Make Policy and Entry relationship unidirectional
This is to fix the error "Referential integrity constraint violation"
that could occur when deleting a LifeCyclePolicy from the database.
Arguably this being a unidirectional relation makes sense, as an
ACL entry does not need to know anything about the policy it belongs to.

Would still be interested in how to solve this properly for bidirectional
relationships.