Workaround for "laggy" servers table in admin ui
Fixes https://github.com/opencast/opencast-admin-interface/issues/689.

Fixes the issue that toggling maintenance mode for a server in the admin ui
is only reflected in the ui after 0 to 60 seconds. The changes should now
be reflected immediately.

What is happening:
The backend is caching all server data it returns to the frontend, updating
it only after at least 60 seconds have passed. This is due to the statistics
queries being expensive enough to potentially take down the database if many
users have the servers table open in the frontend ui. Problem is, all
non-statistics data like maintenance status is cached as well. This patch
removes all non-statistics data from the cache, so that it can always be
up-to-date.