Same target tag behaviour for all WOH
All workflow operation handlers adapt the same method of
handling target tags (if they support target tags).

The chosen behaviour is that of the tag WOH:
"If a target-tag starts with a '-', it will be removed from
preexisting tags, if a target-tag starts with a '+', it will
be added to preexisting tags. If there is no prefix, all
preexisting tags are removed and replaced by the target-tags."

This means that all WOH that did not support "+" and "-" now do.
!THIS CONSTITUTES A BEHAVIOUR CHANGE FOR SOME WORKFLOW OPERATION HANDLERS!
In particular, some WOH were adding tags instead of replacing them,
even though they were not prefaced with a "+". Adopters are advised
to check their workflows.