Use namedQueries instead of ASelectQueries in AssetManagerImpl
Some of the methods in the assetmanager do nothing but get snapshots
based on various arguments. Instead of using querydsl, this adds some
NamedQueries to SnapshotDto that can be used instead.

This is much more wordy than before, mostly due to many slightly different
JPQL queries (acceptable?) and code repetition for auth checking which
could definitely be more concise.