Enable feature "Episode ID Roles" by default
This is a very useful feature to have, especially for external systems.
Especially especially so with JWT, which depends on this to make the
`oc` claim work. For external apps, it's extremely useful to be able
to rely on the presence of this feature and it will be annoying to tell
everyone to enable it manually. For example, Tobira is already making
use of this.