Fix OpenMetrics Endpoint
Pull request #6574 (c89a601) switched from the Prometheus text format to
the OpenMetrics format for our metrics endpoint. Unfortunately, while we
switched the format itself, we didn't change the header announcing the
format type, causing the endpoint to not be compliant with the
OpenMetrics standard and causing e.g. Prometheus to not accept the data.

This patch fixes the problem by announcing the correct content type,
unbreaking the broken metrics endpoint.