Remove entwine JSON
Replace entwine code that handles JSON with Gson.
Gson was chosen because it is a library that we also use.

This mostly affects the admin ui and external api endpoints, since
entwine JSON was used there for responses.

Not sure if this warrants a new external api version, since this
*should* not result in functional changes. Will create a new version
if someone wants me to.