Query Tobira episode/series ID too
If you've configured Tobira correctly, a Tobira tab will appear in the event/series details dialog. This provides a direct link to the resource in Tobira and lists all Tobira realms that contain it. The URLs are based on the Opencast resource ID. Since Tobira uses its own IDs (a hash-like format), we should preferably use these. However, the Tobira resource ID should also be requested.