Deselecting a series may result in loss of access to the event
If the event has no episode ACL but only a series ACL, deselecting a series may result in the event no longer having any ACL. The non admin user will then lose access to the event. This occurs because the series ACL will be applied or (in this case) removed during the metadata update. To prevent this, we can keep the currently active ACL as the episode ACL before deleting the series ACL.