Update Whisper.cpp defaults to reflect new binary names
The default binary names of Whisper.cpp have changed. This patch adjusts
the Opencast code to reflect the new binary names. For more details, see
the deprecation warning at
https://github.com/ggml-org/whisper.cpp/tree/v1.7.5/examples/deprecation-warning

Starting with Opencast 18, this change will also be included in Opencast's
package repository. The additional change to the default model folder is
done to also match the packages.