Remove config `episode.id.role.access` & always enable ID based access
This removes the configuration for this feature, making it always
enabled. Having less variability in these configurations makes it easier
for external apps to rely on Opencast behavior, which in turn makes
integration easier for admins. And this feature in particular will
likely be widely used by external apps via JWT auth.