Faster Tests (Build JavaDocs only once)
This tests build Opencast with two Java versions, but they run the
integration tests only with one version. The JavaDocs are also built
with two versions.

It's somewhat unlikely that the JavaDocs break with a Java version
update, nor do we really care that much. That is why this patch makes
them run only once on the version that is not running the integration
tests, so that they can run in parallel. This should make tests finish
about 5 minutes faster.