Allow to escape colons in workflow variable syntax
Workflow variables have the following syntax:

    ${var:default}

A colon is used to separate the variable name from the default value.
However, the variable itself might contain a colon and there is
currently no way to use such a variable. A specific use case is the
`analyze-mediapackage` operation with `set-tag-variables` enabled. For
subtitle tracks, this might add a workflow property such as

    captions_source_hastag_lang:en

This PR changes the regex to parse variables and allows to escape
colons in the following way:

    ${var-with-\:-colon:default-with\:-colon}

This is breaking in the sense that the string `\:` is now replaced in
the variable name and default value with a `:`, but I don't think
anybody used `\:` at the moment. So this should not be breaking in
practice.