Allow multiple videos for SubtitleTimeShift WOH
Helps with #6895.

The SubtitleTimeshiftWOH only allowed one flavor with exactly one track in it
for video source flavors. This changes it so that it allows multiple flavors
with multiple tracks in it. This is supposed to make it easier to write workflows
where the subtitles have to be shifted by the duration of multiple tracks (e.g.
by a themes title slide AND bumper video).