Reactive user role filter for admin ui
Allows users of the admin interface to filter the users table
by roles once again.

The filter was originally removed due to performance concerns and
to this day it is still not terribly performant. On my local testing
instance with ~20000 users, a request with a role filter could take
between 350~400ms, compared to 80~120ms without a role filter.
However, we are allowing admin interface users to use the textFilter
(aka the search bar) to filter by role, which has the exact same performance
concerns as using the role filter directly. Therefore I argue that we
should either allow users to use the role filter, or disable the
textFilter either just for roles or alltogether.